\name{SparseReporter}
\alias{SparseReporter}
\alias{SparseReporter-class}
\title{Test reporter: sparse}
\description{
  A truncated version of the default 'summary' reporter
  that only prints the failed assertion. Useful as a TODO
  list of un-implemented features if one has otherwise
  passing tests. 
}
\keyword{debugging}

